Moved blob functionality to new repository
