+++ src/Storage/Repository/DataRepository.cs
@@ -1,18 +1,12 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.IO;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 
 using Altinn.Platform.Storage.Configuration;
 using Altinn.Platform.Storage.Interface.Models;
-
-using Azure;
-using Azure.Storage;
-using Azure.Storage.Blobs;
-using Azure.Storage.Blobs.Models;
 
 using Microsoft.Azure.Cosmos;
 using Microsoft.Azure.Cosmos.Linq;
@@ -24,119 +18,29 @@ namespace Altinn.Platform.Storage.Repository
 {
     /// <summary>
     /// Represents an implementation of <see cref="IDataRepository"/> using Azure CosmosDB to keep metadata
-    /// and Azure Blob storage to keep the actual data. Blob storage is again split based on application owner.
     /// </summary>
     internal sealed class DataRepository : BaseRepository, IDataRepository
     {
         private const string CollectionId = "dataElements";
         private const string PartitionKey = "/instanceGuid";
 
-        private readonly AzureStorageConfiguration _storageConfiguration;
-        private readonly ISasTokenProvider _sasTokenProvider;
         private readonly ILogger<DataRepository> _logger;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DataRepository"/> class
         /// </summary>
-        /// <param name="sasTokenProvider">A provider that can be asked for SAS tokens.</param>
         /// <param name="cosmosSettings">the configuration settings for azure cosmos database</param>
-        /// <param name="storageConfiguration">the storage configuration for azure blob storage</param>
         /// <param name="logger">The logger to use when writing to logs.</param>
         /// <param name="cosmosClient">CosmosClient singleton</param>
         public DataRepository(
-            ISasTokenProvider sasTokenProvider,
             IOptions<AzureCosmosSettings> cosmosSettings,
-            IOptions<AzureStorageConfiguration> storageConfiguration,
             ILogger<DataRepository> logger,
             CosmosClient cosmosClient)
             : base(CollectionId, PartitionKey, cosmosSettings, cosmosClient)
         {
-            _storageConfiguration = storageConfiguration.Value;
-            _sasTokenProvider = sasTokenProvider;
             _logger = logger;
         }
-
-        /// <inheritdoc/>
-        public async Task<(long ContentLength, DateTimeOffset LastModified)> WriteDataToStorage(string org, Stream stream, string blobStoragePath)
-        {
-            try
-            {
-                var blobProps = await UploadFromStreamAsync(org, stream, blobStoragePath);
-                return (blobProps.ContentLength, blobProps.LastModified);
-            }
-            catch (RequestFailedException requestFailedException)
-            {
-                switch (requestFailedException.ErrorCode)
-                {
-                    case "AuthenticationFailed":
-                        _logger.LogWarning("Authentication failed. Invalidating SAS token.");
-
-                        _sasTokenProvider.InvalidateSasToken(org);
-
-                        // No use retrying upload as the original stream can't be reset back to start.
-                        throw;
-                    default:
-                        throw;
-                }
-            }
-        }
-
-        /// <inheritdoc/>
-        public async Task<Stream> ReadDataFromStorage(string org, string blobStoragePath)
-        {
-            try
-            {
-                return await DownloadBlobAsync(org, blobStoragePath);
-            }
-            catch (RequestFailedException requestFailedException)
-            {
-                switch (requestFailedException.ErrorCode)
-                {
-                    case "AuthenticationFailed":
-                        _logger.LogWarning("Authentication failed. Invalidating SAS token and retrying download operation.");
-
-                        _sasTokenProvider.InvalidateSasToken(org);
-
-                        return await DownloadBlobAsync(org, blobStoragePath);
-                    case "BlobNotFound":
-                        _logger.LogWarning("Unable to find a blob based on the given information - {org}: {blobStoragePath}", org, blobStoragePath);
-
-                        // Returning null because the blob does not exist.
-                        return null;
-                    case "InvalidRange":
-                        _logger.LogWarning("Found possibly empty blob in storage for {org}: {blobStoragePath}", org, blobStoragePath);
-
-                        // Returning empty stream because the blob does exist, but it is empty.
-                        return new MemoryStream();
-                    default:
-                        throw;
-                }
-            }
-        }
-
-        /// <inheritdoc/>
-        public async Task<bool> DeleteDataInStorage(string org, string blobStoragePath)
-        {
-            try
-            {
-                return await DeleteIfExistsAsync(org, blobStoragePath);
-            }
-            catch (RequestFailedException requestFailedException)
-            {
-                switch (requestFailedException.ErrorCode)
-                {
-                    case "AuthenticationFailed":
-                        _logger.LogWarning("Authentication failed. Invalidating SAS token and retrying delete operation.");
-
-                        _sasTokenProvider.InvalidateSasToken(org);
-
-                        return await DeleteIfExistsAsync(org, blobStoragePath);
-                    default:
-                        throw;
-                }
-            }
-        }
-
+ 
         /// <inheritdoc/>
         public async Task<List<DataElement>> ReadAll(Guid instanceGuid)
         {
@@ -270,65 +174,6 @@ namespace Altinn.Platform.Storage.Repository
             var response = await Container.DeleteItemAsync<DataElement>(dataElement.Id, new PartitionKey(dataElement.InstanceGuid));
 
             return response.StatusCode == HttpStatusCode.NoContent;
-        }
-
-        private async Task<BlobProperties> UploadFromStreamAsync(string org, Stream stream, string fileName)
-        {
-            BlobClient blockBlob = await CreateBlobClient(org, fileName);
-            BlobUploadOptions options = new()
-            {
-                TransferValidation = new UploadTransferValidationOptions { ChecksumAlgorithm = StorageChecksumAlgorithm.MD5 }
-            };
-            await blockBlob.UploadAsync(stream, options);
-            BlobProperties properties = await blockBlob.GetPropertiesAsync();
-
-            return properties;
-        }
-
-        private async Task<Stream> DownloadBlobAsync(string org, string fileName)
-        {
-            BlobClient blockBlob = await CreateBlobClient(org, fileName);
-
-            Azure.Response<BlobDownloadInfo> response = await blockBlob.DownloadAsync();
-
-            return response.Value.Content;
-        }
-
-        private async Task<bool> DeleteIfExistsAsync(string org, string fileName)
-        {
-            BlobClient blockBlob = await CreateBlobClient(org, fileName);
-
-            bool result = await blockBlob.DeleteIfExistsAsync();
-
-            return result;
-        }
-
-        private async Task<BlobClient> CreateBlobClient(string org, string blobName)
-        {
-            if (!_storageConfiguration.AccountName.StartsWith("devstoreaccount1"))
-            {
-                string sasToken = await _sasTokenProvider.GetSasToken(org);
-
-                string accountName = string.Format(_storageConfiguration.OrgStorageAccount, org);
-                string containerName = string.Format(_storageConfiguration.OrgStorageContainer, org);
-
-                UriBuilder fullUri = new()
-                {
-                    Scheme = "https",
-                    Host = $"{accountName}.blob.core.windows.net",
-                    Path = $"{containerName}/{blobName}",
-                    Query = sasToken
-                };
-
-                return new BlobClient(fullUri.Uri);
-            }
-
-            StorageSharedKeyCredential storageCredentials = new(_storageConfiguration.AccountName, _storageConfiguration.AccountKey);
-            Uri storageUrl = new(_storageConfiguration.BlobEndPoint);
-            BlobServiceClient commonBlobClient = new(storageUrl, storageCredentials);
-            BlobContainerClient blobContainerClient = commonBlobClient.GetBlobContainerClient(_storageConfiguration.StorageContainer);
-
-            return blobContainerClient.GetBlobClient(blobName);
         }
     }
 }
