+++ test/UnitTest/Mocks/Repository/DataRepositoryMock.cs
@@ -37,7 +37,7 @@ namespace Altinn.Platform.Storage.UnitTest.Mocks.Repository
             return await Task.FromResult(true);
         }
 
-        public async Task<bool> DeleteDataInStorage(string org, string blobStoragePath)
+        public async Task<bool> DeleteBlob(string org, string blobStoragePath)
         {
             return await Task.FromResult(true);
         }
@@ -109,7 +109,7 @@ namespace Altinn.Platform.Storage.UnitTest.Mocks.Repository
             return await Task.FromResult(dataElements);
         }
 
-        public async Task<Stream> ReadDataFromStorage(string org, string blobStoragePath)
+        public async Task<Stream> ReadBlob(string org, string blobStoragePath)
         {
             string dataPath = Path.Combine(GetDataBlobPath(), blobStoragePath);
             Stream fs = File.OpenRead(dataPath);
@@ -152,7 +152,7 @@ namespace Altinn.Platform.Storage.UnitTest.Mocks.Repository
             return dataElement;
         }
 
-        public async Task<(long ContentLength, DateTimeOffset LastModified)> WriteDataToStorage(string org, Stream stream, string blobStoragePath)
+        public async Task<(long ContentLength, DateTimeOffset LastModified)> WriteBlob(string org, Stream stream, string blobStoragePath)
         {
             MemoryStream memoryStream = new MemoryStream();
             await stream.CopyToAsync(memoryStream);
